Move Buzzer out to its own module.
